+++ src/App.js
@@ -1,6 +1,6 @@
 import "./App.css";
 import {Routes, Route, } from "react-router-dom";
-import {LandingRoute} from './routes/landingroute'
+import {LandingRoute} from './routes/landingRoute'
 
 function App() {
   return (
